Feat: Remove unnecessary dialogue-related code. #10427
